Split test files into common, applicability and expectation
