+++ packages/alfa-rules/tsconfig.json
@@ -190,9 +190,11 @@
     "test/common/predicate/is-at-the-start.spec.tsx",
     "test/common/predicate/is-visible.spec.tsx",
     "test/common/predicate/has-role.spec.tsx",
-    "test/sia-er62/rule.spec.tsx",
-    "test/sia-er87/rule.spec.tsx",
+    "test/sia-er62/rule.applicability.spec.tsx",
+    "test/sia-er62/rule.expectation.spec.tsx",
     "test/sia-er62/serialise.spec.ts",
+    "test/sia-er62/common.ts",
+    "test/sia-er87/rule.spec.tsx",
     "test/sia-r1/rule.spec.tsx",
     "test/sia-r2/rule.spec.tsx",
     "test/sia-r3/rule.spec.tsx",
